# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Olimpiadas",
        "description": "Documentación de la API para evaluaciones y evaluadores",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/api/fases": {
            "get": {
                "tags": [
                    "Fases"
                ],
                "summary": "Listar todas las fases",
                "operationId": "72d4c98364be72fd0c6e605d4ff7de2a",
                "responses": {
                    "200": {
                        "description": "Lista de fases obtenida exitosamente"
                    }
                }
            },
            "post": {
                "tags": [
                    "Fases"
                ],
                "summary": "Registrar una nueva fase",
                "operationId": "50ac039336e75a62851c4024b38cb9e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "estado",
                                    "fecha_inicio",
                                    "fecha_fin"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Fase Eliminatoria"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Primera ronda de evaluaciones"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "example": "abierto"
                                    },
                                    "fecha_inicio": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-01"
                                    },
                                    "fecha_fin": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Fase creada correctamente"
                    },
                    "422": {
                        "description": "Error de validación en los datos enviados"
                    }
                }
            }
        },
        "/api/fases/{id}": {
            "delete": {
                "tags": [
                    "Fases"
                ],
                "summary": "Eliminar una fase existente",
                "operationId": "795f170a1cc69cf038b0a369bceb7b91",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la fase a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fase eliminada correctamente"
                    },
                    "404": {
                        "description": "Fase no encontrada"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Fases"
                ],
                "summary": "Actualizar parcialmente una fase",
                "operationId": "3d00cf2cd32bf1baf4b26486e1382553",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la fase a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Fase Final"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Última etapa de la competencia"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "example": "cerrado"
                                    },
                                    "fecha_inicio": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-20"
                                    },
                                    "fecha_fin": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-30"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Fase actualizada correctamente"
                    },
                    "404": {
                        "description": "Fase no encontrada"
                    }
                }
            }
        },
...
    },
...
}